requirements.txt
# === Core Python Utilities ===
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.2.1

# === FastAPI & Server (Optional but recommended) ===
fastapi==0.115.0
uvicorn==0.30.0

# === Database ===
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# === NLP / Extractors ===
spacy==3.7.4
spacy-langdetect
# ==0.1.0
beautifulsoup4==4.12.3
lxml==5.2.2

# === Embeddings / Similarity ===
sentence-transformers==2.7.0

# === LLM Clients  ===
# openai==1.35.14
# google-generativeai==0.6.1
groq==0.10.0

# === PDF Generation ===
reportlab==4.1.0
weasyprint==62.3

# === Utilities ===
requests==2.32.3
httpx==0.27.0
tqdm==4.66.4

# === Testing ===
pytest==8.2.0
pytest-asyncio==0.23.7

